requirements.txt
gdown==5.2.0
numpy==2.1.3
Pillow==11.0.0
streamlit==1.40.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
